refactor: break down index.tsx into modular components and hooks
- Created useMessageHandling hook for all message API logic
- Created useFileUpload hook for file upload and drag & drop
- Created useTeamConfig hook for team configuration management
- Created utility files: forms.ts, keyboard.ts for reusable logic
- Created new components: ChatContainer, DragDropOverlay, AppLayout
- Reduced index.tsx from 1548 lines to ~300 lines
- Maintained all existing functionality while improving code organization
- All tests passing, build successful
