+++ src/components/ChatContainer.tsx
@@ -1,0 +1,168 @@
+import { FunctionComponent } from 'preact';
+import { useState, useEffect, useRef } from 'preact/hooks';
+import VirtualMessageList from './VirtualMessageList';
+import MessageComposer from './MessageComposer';
+import { ChatMessageUI } from '../../worker/types';
+import { FileAttachment } from '../../worker/types';
+import { createKeyPressHandler } from '../utils/keyboard';
+
+interface ChatContainerProps {
+  messages: ChatMessageUI[];
+  onSendMessage: (message: string, attachments: FileAttachment[]) => void;
+  onDateSelect?: (date: Date) => void;
+  onTimeOfDaySelect?: (timeOfDay: 'morning' | 'afternoon') => void;
+  onTimeSlotSelect?: (timeSlot: Date) => void;
+  onRequestMoreDates?: () => void;
+  onServiceSelect?: (service: string) => void;
+  onUrgencySelect?: (urgency: string) => void;
+  onCreateMatter?: () => void;
+  onScheduleConsultation?: () => void;
+  onLearnServices?: () => void;
+  teamConfig?: {
+    name: string;
+    profileImage: string | null;
+    teamId: string;
+    description?: string | null;
+  };
+  onOpenSidebar?: () => void;
+  sessionId?: string;
+  teamId?: string;
+  onFeedbackSubmit?: (feedback: any) => void;
+  
+  // File handling props
+  previewFiles: FileAttachment[];
+  removePreviewFile: (index: number) => void;
+  handlePhotoSelect: (files: File[]) => Promise<void>;
+  handleCameraCapture: (file: File) => Promise<void>;
+  handleFileSelect: (files: File[]) => Promise<void>;
+  handleMediaCapture: (blob: Blob, type: 'audio' | 'video') => void;
+  isRecording: boolean;
+  setIsRecording: (v: boolean) => void;
+}
+
+const ChatContainer: FunctionComponent<ChatContainerProps> = ({
+  messages,
+  onSendMessage,
+  onDateSelect,
+  onTimeOfDaySelect,
+  onTimeSlotSelect,
+  onRequestMoreDates,
+  onServiceSelect,
+  onUrgencySelect,
+  onCreateMatter,
+  onScheduleConsultation,
+  onLearnServices,
+  teamConfig,
+  onOpenSidebar,
+  sessionId,
+  teamId,
+  onFeedbackSubmit,
+  previewFiles,
+  removePreviewFile,
+  handlePhotoSelect,
+  handleCameraCapture,
+  handleFileSelect,
+  handleMediaCapture,
+  isRecording,
+  setIsRecording
+}) => {
+  const [inputValue, setInputValue] = useState('');
+  const textareaRef = useRef<HTMLTextAreaElement>(null);
+
+  const handleInputChange = (e: Event) => {
+    const target = e.currentTarget as HTMLTextAreaElement;
+    setInputValue(target.value);
+    
+    // Simple approach: reset height then set to scrollHeight
+    target.style.height = '24px'; // Reset to default height first
+    target.style.height = `${Math.max(24, target.scrollHeight)}px`;
+  };
+
+  // Simple resize handler for window size changes
+  useEffect(() => {
+    const handleResize = () => {
+      const textarea = document.querySelector('.message-input') as HTMLTextAreaElement;
+      if (textarea) {
+        // Use the same improved auto-expand logic
+        textarea.style.height = '0';
+        const newHeight = Math.max(24, textarea.scrollHeight);
+        textarea.style.height = `${newHeight}px`;
+      }
+    };
+
+    window.addEventListener('resize', handleResize);
+    return () => window.removeEventListener('resize', handleResize);
+  }, []);
+
+  // Initialize textarea height on mount
+  useEffect(() => {
+    const textarea = document.querySelector('.message-input') as HTMLTextAreaElement;
+    if (textarea && textarea.value) {
+      textarea.style.height = '0';
+      const newHeight = Math.max(24, textarea.scrollHeight);
+      textarea.style.height = `${newHeight}px`;
+    }
+  }, []);
+
+  const handleSubmit = () => {
+    if (!inputValue.trim() && previewFiles.length === 0) return;
+
+    const message = inputValue.trim();
+    const attachments = [...previewFiles];
+    
+    // Send message to API
+    onSendMessage(message, attachments);
+    
+    // Reset input and focus
+    setInputValue('');
+    
+    // Just focus the textarea
+    const textarea = document.querySelector('.message-input') as HTMLTextAreaElement;
+    if (textarea) {
+      textarea.focus();
+    }
+  };
+
+  const handleKeyPress = createKeyPressHandler(handleSubmit);
+
+  return (
+    <div className="flex flex-col h-screen w-full m-0 p-0 relative overflow-hidden bg-white dark:bg-dark-bg pt-16 lg:pt-0">
+      <main className="flex flex-col h-full w-full overflow-hidden relative bg-white dark:bg-dark-bg">
+        <VirtualMessageList
+          messages={messages}
+          onDateSelect={onDateSelect}
+          onTimeOfDaySelect={onTimeOfDaySelect}
+          onTimeSlotSelect={onTimeSlotSelect}
+          onRequestMoreDates={onRequestMoreDates}
+          onServiceSelect={onServiceSelect}
+          onUrgencySelect={onUrgencySelect}
+          onCreateMatter={onCreateMatter}
+          onScheduleConsultation={onScheduleConsultation}
+          onLearnServices={onLearnServices}
+          teamConfig={teamConfig}
+          onOpenSidebar={onOpenSidebar}
+          sessionId={sessionId}
+          teamId={teamId}
+          onFeedbackSubmit={onFeedbackSubmit}
+        />
+        <MessageComposer
+          inputValue={inputValue}
+          setInputValue={setInputValue}
+          previewFiles={previewFiles}
+          removePreviewFile={removePreviewFile}
+          handlePhotoSelect={handlePhotoSelect}
+          handleCameraCapture={handleCameraCapture}
+          handleFileSelect={handleFileSelect}
+          handleScheduleStart={onScheduleConsultation}
+          isRecording={isRecording}
+          handleMediaCapture={handleMediaCapture}
+          setIsRecording={setIsRecording}
+          onSubmit={handleSubmit}
+          onKeyPress={handleKeyPress as any}
+        />
+      </main>
+    </div>
+  );
+};
+
+export default ChatContainer;
